requirements.txt
asgiref
certifi
charset-normalizer
Django
idna
pillow
psycopg2-binary
requests
sqlparse
typing_extensions
urllib3
ollama
